begin work on new mappers
